+++ client/src/components/page-header.tsx
@@ -1,19 +1,23 @@
 import { Flex, FlexProps, Heading, Text } from '@chakra-ui/react';
-import SyncSpinner from './spinners/sync-spinner';
+import LoadingStatus from './loading/loading-status';
 
 interface PageHeaderProps extends FlexProps {
   action?: React.ReactNode;
   isFetching?: boolean;
+  isRefreshing?: boolean;
   subHeader?: string;
 }
 
-const PageHeader: React.FC<PageHeaderProps> = ({ action, isFetching, children, subHeader, ...props }) => {
+const PageHeader: React.FC<PageHeaderProps> = ({ action, isFetching, isRefreshing, children, subHeader, ...props }) => {
+  const showRefreshing = isRefreshing ?? isFetching;
+
   return (
     <Flex justifyContent="space-between" direction="column" gap={2} mb="4" {...props}>
       <Flex justify="space-between" align="start" gap="3">
-        <Heading textStyle="pageTitle">
-          {children} {isFetching && <SyncSpinner size="sm" />}
-        </Heading>
+        <Flex align="center" gap="2">
+          <Heading textStyle="pageTitle">{children}</Heading>
+          {showRefreshing && <LoadingStatus />}
+        </Flex>
         {action}
       </Flex>